Bug fixes, new setting, code clean up, and added support u64::max values per arg

This turned into a little more than just a quick bug fix :stuck_out_tongue_winking_eye:

Thanks to #409 I've piggy backed on that and added support for > 256 values as well.

There's also just a bunch of code deduping and super small macro cleanups.

After this merges I'll put out 2.0.3

+++ src/args/any_arg.rs
@@ -12,9 +12,9 @@ pub trait AnyArg<'n, 'e>: Display {
     fn is_set(&self, ArgSettings) -> bool;
     fn set(&mut self, ArgSettings);
     fn has_switch(&self) -> bool;
-    fn max_vals(&self) -> Option<u8>;
-    fn min_vals(&self) -> Option<u8>;
-    fn num_vals(&self) -> Option<u8>;
+    fn max_vals(&self) -> Option<u64>;
+    fn min_vals(&self) -> Option<u64>;
+    fn num_vals(&self) -> Option<u64>;
     fn possible_vals(&self) -> Option<&[&'e str]>;
     fn validator(&self) -> Option<&Rc<Fn(String) -> Result<(), String>>>;
     fn short(&self) -> Option<char>;
